refactor: reorganize infinite-canvas with clean folder structure
New structure:
- components/          # React Three Fiber components
  - MediaPlane.tsx     # Individual image plane component
  - Chunk.tsx          # Chunk group component
- lib/                 # Utilities and configuration
  - types.ts           # TypeScript types
  - constants.ts       # Canvas constants
  - math-utils.ts      # Math helper functions
  - chunk-generator.ts # Chunk/plane generation logic
  - texture-loader.ts  # Texture loading (renamed from manager)
- styles/              # CSS modules
  - infinite-canvas.module.css
- CanvasScene.tsx      # Main Three.js canvas scene
- InfiniteCanvas.tsx   # Lazy-loaded wrapper component
- index.ts             # Public exports

Changes:
- Move manifest.json to src/data/artworks-manifest.json
- Split scene.tsx into CanvasScene.tsx + components
- Rename texture-manager.ts to texture-loader.ts
- Organize utilities into logical files
- Update all imports to use new structure
