+++ frontend/src/components/sections/infinite-canvas/index.ts
@@ -1,0 +1,2 @@
+export { InfiniteCanvas } from "./InfiniteCanvas";
+export type { InfiniteCanvasProps, MediaItem } from "./lib/types";
