+++ frontend/src/components/sections/infinite-canvas/lib/math-utils.ts
@@ -1,0 +1,23 @@
+export const clamp = (v: number, min: number, max: number): number =>
+    Math.max(min, Math.min(max, v));
+
+export const lerp = (a: number, b: number, t: number): number => a + (b - a) * t;
+
+export const seededRandom = (seed: number): number => {
+    const x = Math.sin(seed * 9999) * 10000;
+    return x - Math.floor(x);
+};
+
+export const hashString = (str: string): number => {
+    let h = 0;
+    for (let i = 0; i < str.length; i++) h = ((h << 5) - h + str.charCodeAt(i)) | 0;
+    return Math.abs(h);
+};
+
+export const shouldThrottleUpdate = (
+    lastUpdateTime: number,
+    throttleMs: number,
+    currentTime: number
+): boolean => {
+    return currentTime - lastUpdateTime >= throttleMs;
+};
