+++ frontend/src/components/sections/infinite-canvas/styles/infinite-canvas.module.css
@@ -1,0 +1,24 @@
+/** biome-ignore-all lint/complexity/noImportantStyles: canvas needs explicit sizing */
+.container {
+    width: 100%;
+    height: 100%;
+    position: absolute;
+    inset: 0;
+    touch-action: none;
+}
+
+.canvas {
+    background-color: transparent;
+    width: 100%;
+    height: 100%;
+    position: absolute;
+    inset: 0;
+    touch-action: none;
+}
+
+.stats {
+    position: absolute !important;
+    top: 12px !important;
+    right: 12px !important;
+    left: auto !important;
+}
